pyproject.toml
[project]
name = "cricket-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.3.2",
    "ipykernel>=6.29.5",
    "numpy>=2.3.1",
    "orjson>=3.11.0",
    "pandas>=2.3.1",
    "polars>=1.31.0",
    "pyarrow>=21.0.0",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]

[project.scripts]
cricket-etl = "cricket_etl:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["src"]
